Move parser from ir.parser to language.parser namespace
Relocates parser components from pypto.ir.parser to pypto.language.parser
to better reflect architectural separation between the high-level language
DSL and the IR layer. Updates all imports, documentation, and tests to
reference the new namespace.
