+++ python/pypto/__init__.py
@@ -16,7 +16,7 @@ This package provides Python bindings for the PyPTO C++ library.
 from typing import cast
 
 # Import IR module (includes operations and core IR types)
-from . import ir
+from . import ir, language
 from .pypto_core import (
     DataType,
     InternalError,
@@ -57,6 +57,7 @@ __all__ = [
     # Modules
     "testing",
     "ir",
+    "language",
     # Logging framework
     "InternalError",
     "LogLevel",
