+++ python/pypto/language/parser/scope_manager.py
